+++ ai_hydra/nnet/Policy/LinearPolicy.py
@@ -8,29 +8,10 @@
 #    License: GPL 3.0
 
 import torch
-import torch.nn as nn
 from typing import Sequence
 
-from ai_hydra.nnet.HydraPolicy import HydraPolicy
-from ai_hydra.constants.DNet import DNetDef, DLinear
-from ai_hydra.constants.DHydra import DHydra
-
-
-class LinearModel(nn.Module):
-    def __init__(self):
-        super().__init__()
-        torch.manual_seed(DHydra.RANDOM_SEED)
-
-        self.net = nn.Sequential(
-            nn.Linear(DNetDef.INPUT_SIZE, DLinear.HIDDEN_SIZE),
-            nn.ReLU(),
-            nn.Linear(DLinear.HIDDEN_SIZE, DLinear.HIDDEN_SIZE),
-            nn.ReLU(),
-            nn.Linear(DLinear.HIDDEN_SIZE, DLinear.OUTPUT_SIZE),
-        )
-
-    def forward(self, x):
-        return self.net(x)
+from ai_hydra.nnet.Policy.HydraPolicy import HydraPolicy
+from ai_hydra.nnet.models import LinearModel
 
 
 class LinearPolicy(HydraPolicy):