+++ ai_hydra/nnet/RandomPolicy.py
@@ -10,7 +10,7 @@
 import random
 from typing import Sequence
 
-from ai_hydra.nnet.HydraPolicy import HydraPolicy
+from ai_hydra.nnet.Policy.HydraPolicy import HydraPolicy
 
 
 class RandomPolicy(HydraPolicy):