+++ ai_hydra/nnet/TrainMgr.py
@@ -6,7 +6,7 @@ from typing import Optional
 from ai_hydra.server.SnakeMgr import SnakeMgr
 from ai_hydra.constants.DGame import DGameField
 
-from ai_hydra.nnet.HydraPolicy import HydraPolicy
+from ai_hydra.nnet.Policy.HydraPolicy import HydraPolicy
 from ai_hydra.nnet.Trainer import Trainer
 from ai_hydra.nnet.ReplayMemory import ReplayMemory
 from ai_hydra.nnet.Transition import Transition